xray: test demonstrating emit failure for subsegment with closed parent
This isn't necessarily a realistic use case, but it is possible to set
up this situation due to the package interface. Should this happen,
the segment will quietly fail to flush the trace without any warning.

It seems to be that `Segment.openSegments` will become negative
resulting in `Segment.flush` ignoring the segment because it will only
handle a segment with exactly _zero_ `openSegments`.